requirement.txt
pillow
numpy
tensorflow
opencv-python